Random CTE

+++ customs.cpp
@@ -7,7 +7,6 @@
 #include <set>
 #include <stack>
 #include <string.h>
-#include <math.h>
 
 using namespace std;
 
@@ -79,7 +78,7 @@ int main(){
     }
     bool flag = true; long long prod = 1ll;
     for(int i = 0; i < components.size(); i++){
-        prod = (prod*components[i].size())%MOD;
+        prod = (prod*(long long)components[i].size())%MOD;
         if(!isLoop(components[i])){ flag = false; break; }
     }
     if(!flag){ cout << "safe\n"; return 0; }
